debug

+++ src/main/scala/edu/berkeley/cs/rise/opaque/execution/operators.scala
@@ -71,7 +71,7 @@ case class EncryptedLocalTableScanExec(
   override def executeBlocked(): RDD[Block] = {
     // Locally partition plaintextData using the same logic as ParallelCollectionRDD.slice
     def positions(length: Long, numSlices: Int): Iterator[(Int, Int)] = {
-      (0 until length).iterator.map { i =>
+      (0 until length.toInt).iterator.map { i =>
 //        val start = ((i * length) / numSlices).toInt
 //        val end = (((i + 1) * length) / numSlices).toInt
 //        (start, end)
